Dec update
This was a lingering update that I never got around to finishing. The only thing missing was scraping the completion dates from the registries. Rather than do that, I've just put all the missing ones as "Not Available" for now. If I get around to another update, I will do a more recent scrape to fill in the holes.